chore: divide TusController component to another file
